+++ packages/dataviews/src/components/dataviews-bulk-actions-toolbar/style.scss
@@ -1,45 +1,5 @@
-.dataviews-bulk-actions-toolbar {
-	position: sticky;
+.dataviews-bulk-actions-toolbar__selection-count {
 	display: flex;
-	flex-direction: column;
-	align-content: center;
-	flex-wrap: wrap;
-	width: fit-content;
-	margin-left: auto;
-	margin-right: auto;
-	bottom: $grid-unit-30;
-	z-index: z-index(".dataviews-bulk-actions-toolbar");
-
-	.components-accessible-toolbar {
-		border-color: $gray-300;
-		box-shadow: $elevation-x-small;
-
-		.components-toolbar-group {
-			border-color: $gray-200;
-
-			&:last-child {
-				border: 0;
-			}
-		}
-	}
-
-	.dataviews-bulk-actions-toolbar__selection-count {
-		display: flex;
-		align-items: center;
-		margin: 0 $grid-unit-10 0 $grid-unit-10;
-	}
-}
-
-.dataviews-bulk-actions-toolbar__wrapper {
-	display: flex;
-	flex-grow: 1;
-	width: 100%;
-
-	.components-toolbar-group {
-		align-items: center;
-	}
-
-	.components-button.is-busy {
-		max-height: $button-size;
-	}
+	align-items: center;
+	margin: 0 $grid-unit-10 0 $grid-unit-10;
 }
