+++ packages/dataviews/src/components/dataviews-footer/style.scss
@@ -1,4 +1,4 @@
-.dataviews-pagination {
+.dataviews-footer {
 	position: sticky;
 	bottom: 0;
 	left: 0;
@@ -11,15 +11,9 @@
 	@include reduce-motion("transition");
 }
 
-.dataviews-pagination__page-select {
-	font-size: 11px;
-	font-weight: 500;
-	text-transform: uppercase;
-
-	@include break-small() {
-		.components-select-control__input {
-			font-size: 11px !important;
-			font-weight: 500;
-		}
+/* stylelint-disable-next-line scss/at-rule-no-unknown -- '@container' not globally permitted */
+@container (max-width: 430px) {
+	.dataviews-footer {
+		padding: $grid-unit-15 $grid-unit-30;
 	}
 }
