+++ packages/dataviews/src/components/dataviews/index.tsx
@@ -13,7 +13,6 @@ import { useMemo, useState } from '@wordpress/element';
  * Internal dependencies
  */
 import { default as DataViewsBulkActions } from '../dataviews-bulk-actions';
-import DataViewsBulkActionsToolbar from '../dataviews-bulk-actions-toolbar';
 import DataViewsContext from '../dataviews-context';
 import {
 	default as DataViewsFilters,
@@ -21,7 +20,7 @@ import {
 	FilterVisibilityToggle,
 } from '../dataviews-filters';
 import DataViewsLayout from '../dataviews-layout';
-import DataviewsPagination from '../dataviews-pagination';
+import DataViewsFooter from '../dataviews-footer';
 import DataViewsSearch from '../dataviews-search';
 import DataViewsViewConfig from '../dataviews-view-config';
 import { normalizeFields } from '../../normalize-fields';
@@ -153,8 +152,7 @@ export default function DataViews< Item >( {
 				</HStack>
 				{ isShowingFilter && <DataViewsFilters /> }
 				<DataViewsLayout />
-				<DataviewsPagination />
-				<DataViewsBulkActionsToolbar />
+				<DataViewsFooter />
 			</div>
 		</DataViewsContext.Provider>
 	);
