+++ packages/dataviews/src/style.scss
@@ -2,6 +2,7 @@
 @import "./components/dataviews-bulk-actions/style.scss";
 @import "./components/dataviews-bulk-actions-toolbar/style.scss";
 @import "./components/dataviews-filters/style.scss";
+@import "./components/dataviews-footer/style.scss";
 @import "./components/dataviews-pagination/style.scss";
 @import "./components/dataviews-item-actions/style.scss";
 @import "./components/dataviews-selection-checkbox/style.scss";
